add an option to configure 24bit output
create a datatype `ColorMode`, which is stored in `Config`
and `Output`.  It is passed to all of the functions that handle color
processing/output.

Currently this only has one possible value, which gives us the default
vty color handling behavior.

Factor out color drawing code in TerminfoBased.hs

Add datatype for 24-bit colors

Add 24-bit color support

we just write the colors straight to the terminal without using
terminfo.

Move all color-handling info to `ColorMode`

- Add terms for 1/8/16/240 colors to `ColorMode`
- Add function `detectColorMode` to detect color mode

This removes all color-handling information outside of the `ColorMode`
datatype, giving us a single source for color-support information.

This emulates the old color-handling code perfectly pretty much perfectly,
with one exception; For users with >256 colors listed, we don't do anything
special when outputting `Color240`s. The old behavior was absolutely
baffling to me, and seemed like it might cause future issues, so I just
removed it.

Add functions for creating 24-bit colors

- linearColor is just a polymorphic synonym for `RGBColor`
- srgbColor converts inputs from to linear before making a `RGBColor`
- color240 is the same as rgbColor, but with a less confusing name

This is all based of of the assumption that outputs will be linear,
which is true for gnome-terminal, but I haven't tested it with anything
else